+++ Baboon.Angular.App/src/app.module.js
@@ -9,7 +9,8 @@
     'templates',
     'angular-baboon.common',
     'angular-baboon.reports',
-    'angular-baboon.home'
+    'angular-baboon.home',
+    'angular-baboon.login'
   ]);
 
 })();