+++ Baboon.Angular.App/src/modules/login/login.module.js
@@ -1,0 +1,11 @@
+(function () {
+  'use strict';
+
+  angular.module('angular-baboon.login', [
+    'angular-baboon.login.controller'
+
+
+  ]);
+
+
+})();